requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
pillow
python-multipart
jinja2
gradio
pillow